Enhance: show help text when root icon selected in editor
Clicking the root icon should not show any editing window as the root icon
does not represent something that the user can edit.  This commit makes
the help text show as it would if there was NO other items in the editor
tree of the selected type (Trigger, Alias etc,)  As we pre-populate the
editor with one or two items automagically some of the help texts were not
previously normally shown at all!

Testing revealed that it was also necessary to protect the folders that
represent modules - they cannot/should not contain any user enterable
scripts, ect. as they are generated automatically and named at the time
of creation and thus should not be allowed to be changed later as that
change is not savable and could be confusing.  They also need protection
from deletion but that I think is a subject for a separate commit.

In fixing the handling of module folders I was also able to correct a long
standing issue that groups of "Actions" (a.k.a "Buttons") that are "top"
level folders were being identified as Menus and not Toolbars and the wrong
ActionsMainArea widget was being displayed for them if they were included
in a module.

Whilst refactoring the relevant methods I also took the opportunity to:
* optimise various for(...) loops
* remove some redundant intermediate variables that are only used one in a
  method
* some duplication or redundant lines that set something, only for a
  different value to be assigned to the same item in the same block of
  code.
* add some missing '{''}'s around single lines of code
* wrap some QStrings in tr(...) or QStringLiteral(...) as appropriate,
  also revise some tests involving string literals.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>