BugFix: correct module disablement of toolbar/menu/buttons
Testing revealed that disabled toolbars/menus/buttons were not being
hidden/disabled if they were in a module and the module was disabled.
This commit restores the expected action which was to hide "Button" type
entities if they were disabled in the editor especially if they were
contained in a module.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>