Transitioned to a custom postMessage for echoing
Moved from a handful of printDebug statements to instead a single
postMessage making use of a custom color scheme for "[  LUA  ]"
messages, inspired by the appearance of the error in the error view in
the editor dialog.

Also added copyright tags for any files that are edited by this branch.