BugFix: improve concurrent access to HostManager::mHostPool
During testing it was found that it was not possible to raise another event
whilst one was in progress - it might be that this is specific to this new
interprofile event.  It looks as though there may be issues with regard to
multiple things trying to access HostManager::mHostPool at the same time
which was being protected by a QMutex used with a QMutexLocker.  This
commit changes the methods in HostManager to permit shared read access to
the mHostPool which should eliminate this issue as exclusive, WRITE access
is only needed briefy to change that QMap which only happens during profile
loading or closing.

Examination of the source code revealed that some methods in the
HostManager class were not being used - especially when other dead code in
the TLuaInterpreter and dlgConnectionProfiles classes were also commented
out.

Following comparison of the behaviour of the
TLuaInterpreter::raiseEvent(...) function to the new function that is the
primary subject of this series of commits I decided to rename the latter
FROM: "raiseInterProfileEvent" TO: "sendGlobalEvent"
This is because the raiseEvent() command can generate an event with ANY
name (and the user will need to register one or more handlers to process
whatever that name WAS) whereas the new command creates an event that goes
to ALL profiles BUT HAS THE FIXED NAME of "sysInterProfileEvent" and needs
a handler to respond to that PARTICULAR name.  By making the name less
similar I hope to reduce confusion in users/script writers, besides which
it is a bit shorter!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>