Refactor: editor code around activate/disactive item messages
The existing information messages are a bit confusing in that if there is
an error in a script element and the user tries to "deactivate" the item
with it they get a message that warns them that the item could not be
"deactivated" and gives the error message.  This commit reorders the
relevant methods so that such messages are not built up from assembling
strings to the same extent - this is also important for future I18n which
does not work well with such techniques!  It also allows the use of the
same QIcon variable for both functional and error cases.

It also transpires that the (void)dlgTriggerEditor::slot_XXXX_active()
methods are not being used as SLOTs so they can be converted to normal
(public) methods.

During testing it became rather clear that Triggers, Aliases and Keys but
NOT Timers, Scripts and Actions (Buttons) automatically greyed out their
child items in the editor tree display if they themselves (or any parent)
is disabled.  To achieve the same uniformity of this greying out it was
necessary to add calls to "pT->ancestorActive()" for many of the lines
of code that set a particular icon for each item in the Tree display and
to add new methods:
(void) dlgTriggerEditor::children_icon_XXXX(QTreeWidgetItem *)
where XXXX is timer, script or action to recurse through those item types
like their namesakes do for the other (non-Variable) items.  It was
also possible to refactor the code that automatically enabled an item if
it was a "New" one {which was needed as the prior code would have
interfered with the code to grey out items if one of their ancestors was
disabled!}

Also:
* Remove a redundant line from:
  (void) dlgTriggerEditor::slot_debug_mode()
* In (void) dlgTriggerEditor::children_icon_XXXX(QTreeWidgetItem *) method:
  convert a "return:" that would abort a loop through the children
  concerned to a "continue;" that would only skip a child that could not be
  found.
* Refactor dlgTriggerEditor::show_var(...) so it takes a boolean argument
  to control repopulating the variables so the code that is common between
  it and dlgTriggerEditor::slot_show_var() can be reduced to be in it
  alone.
* Refactor all dlgTriggerEditor::slot_show_XXXX() methods so they show the
  information message under all cases where they do not show any elements
  instead of a root element which has at least one child.
* Tweaked the "System Message Area" of the Trigger Editor to align itself
  to the top of it's containing widget so it does not jump around as
  widgets are shown or hidden below it.
* Removed some duplicate QSizePolicy instances in the dlgTriggerEditor
  constructor.
* Refactored TTreeWidget to remove some "goto"s and insert some missing
  initialisers!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>