Revise: rename to raiseGlobalEvent, don't sent event to sender
Following extended discussions it was decided to make the new command more
like the existing raiseEvent() so that it could have any name as the first
argument which becomes the "eventName".

This being the case then it needs (I think) to not send the event generated
to the profile that creates it which is what other parties to the
discussion felt.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>