Revise: lock module master folders in Editor TreeWidget
It was possible to drag the brown folders that represent modules and this
could produce undefined results as they are not normal items that can be
edited (or should be relocated to other places in the tree).

Also remove some pointless temporary QIcon and QString variables nearby to
where the above change was done.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>