Enhance: restore version check to Mudlet XMLimport code
I foresee the need to tweak the Xml format in the future, however though
there once was some checks for this in the code in the past it was
commented out.  This commit - which should go into both development and
release_30 branches despite it not being technically a "BugFix" will now
alert the user if they load any code which does not have a "version"
attribute with a numeric value equivalent to 1.000f in the "MudletPackage"
element of the Xml.

It will not STOP the code being loaded, just issue the alert!

This is to cover the CURRENT code against problems that might arise in the
future if the XML format/structure gets changed which will otherwise go
*undetected* by the Mudlet application as it is currently.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>