BugFix: bad parsing of map room/exit sizes + tidy XML import/export classes
Floating point numbers used for the 2D map room and exit-line sizes were
included in the XML file format twice and being read as integers when one,
at least, is a decimal number in range 0.1 to 1.1 so was being set to zero
when read for 9 out 11 cases.

The miss-spelt duplicate "commandSeperator" element is no longer written as
the parsing code can handle an element missing when reading the Host part
of the XML file.

The XMLimport class will silently ignore the above duplicate details so as
to avoid producing qDebug() messages about them when reading them in XML
files produced from older code.

I have taken the opportunity to clean up the code layout in these two
classes, the XMLexport had some code block indentation issues and the
XMLimport class some pointless if() code.  I also have added in error
detection/reporting code in the export code so that if errors arise in
method function calls they are reported up the caller chain and the
XMLWriter aborts as soon as an error is detected.  I have also added
QDebug() reporting for all instances where lua scripts are inserting into
the Mudlet items as they are loaded where they were only previously being
done for some types of item.

It was also possible to eliminate the redundant (QString) XMLexport::mType
member and the unneeded XMLexport(Host *, bool) C'tor.

I also renamed:
* (void) XMLimport::readMapList(QMap<QString, QStringList> & )
 ==> (void) XMLimport::readModulesDetailsMap(QMap<QString, QStringList> & )

to reduce any confusion over what it applies to (a QMap not the profile's
Map!)

I have held off on applying QStringLiteral(...) wrappers around all the
string literals in these classes as that would just make the change-set
even larger on this occasion but it is something to be done at a later
stage IMHO.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>