Fixup: correct search+replace error and a further logic issue
When doing a search and replace for "pT" in the XMLexport.cpp file I hit a
few cases where that text was a part of a longer variable name and broke
them!

Whilst fixing that I spotted that a boolean flag in several methods in that
class that was intended to be cleared on error was capable of being
returned to true afterwards (although practically it was perhaps unlikely
to happen) - for predictable behaviour the code has been amended to never
allow the flag to be returned to the set state.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>