Bugfix: remove Spam output from Lua [c|d|h]echo commands
I was trying to determine why my script was spamming the Central Debug
Console with lines associate with the use of:
TConsole::select(const QString &, int)
at first I thought it was because we register TLuaInterpreter::select() as
the user Lua command: "selectString" - to avoid any confusion with the
built-in lua command of the same name that is used with variable length
argument lists (often represented by three ... and meaning "all the
remaining arguments") I have taken the opportunity to bring the function
up to current style for error messages and argument handling (it will take
three arguments where the first is an empty string or the translated
equivalent to "main" to behave the same as the two argument case which acts
on the main profile console window.)

This did not "cure" my original concern and then I detected that it was
coming about with the use of "cecho(...)", "decho(...)" and "hecho(...)"
but it was happening inside the Mudlet Lua that is external to the C++
core.  It transpires that the "deselect(...)" command, although defined
in the Core was being by-passed/replaced by Lua code in the GUIUtils.lua
file which although documented as:
"Clears the current selection in the main window or miniConsole window.
 (Note: deselect(windowName) is implemented in Core Mudlet.)"
was being used instead - possibly all the time - to call
'selectString("", 1)' which was what was producing the spam I was seeing.

I have refactored the Core deselect([<windowName>]) and the related
reset([<windowName>]) functions to also bring them up to current standards.
The internal mudlet::resetFormat(...) command that handles the latter was
returning an unused true/false value to indicate whether the given
sub-window/console was found or not and the related
mudlet::resetFormat(...) that handles the former did not - so I have revise
that and both TLuaInterpreter class intermediates to now also return a
single boolean value on success.  For the main console which uses different
direct TConsole class methods for these actions the command will always
succeed so a hard-code true value will be returned for those.

Note that deselect() is now entirely commented out in the external Lua
file GUIUtils.lua but this change will also have to be carried out in the
separate Mudlet-Lua repository holding the "production" version of this
file.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>