BugFix: revise & add colour definitions to provided table
The colour codes predefined in the GUIUtils.lua file that is automatically
included into the Mudlet application uses a set of colour names that seem
to be ALMOST the same as:
https://www.w3.org/TR/SVG11/types.html#ColorKeywords

However there are some omissions and differences:
* these are missing:
  * "dark gray"
  * "silver" (very close to previous value of redefined "gray")
  * "dark blue"
  * "aqua" (the same as "cyan")
  * "teal"
  * "dark cyan"
  * "lime" (the same as the previous value of redefined "green")
  * "light_green"
  * "olive"
  * "fuchsia" (the same as "magenta")
  * "dark magenta"
  * "indigo"

* these are redefined:
  * "gray"/"grey"
  * "green"
  * "maroon" (now the same as "red")
  * "purple"

* these are extra:
  * "light goldenrod"
  * "light slate blue"
  * "navy blue"
  * "violet red"

By default this commit will not alter things for previous scripts/packages
however using the NEW lua command forceWebColorTable( true ) will revise
the table appropriately or should this break things (a script uses a colour
name that was deleted) forceWebColorTable( false ) will revert to the
default table that Mudlet used in the past.

In passing I also noticed a bug in the code that displays the colour table
which meant that there may be some sample colours which have a white text
overlay when it should have been black or visa-versa.

I have also reordered the contents of the color_table to be alphabetical
so that it is easier to find a colour definition.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>