BugFix: make searches/checks for "xml" extension files not case sensitive
This is likely the cause of bug:
https://bugs.launchpad.net/mudlet/+bug/1417234
which is where a suitable file is having an extra xml extension added,
probably because the MacOS File system and native file dialogues are not
necessarily case sensitive and could produce a file with an extension XML
rather than xml that would fail checks for the latter in the past.

This bug regarded as a "ShowStopper" for Mudlet 3.0 release; this fix
ported from that branch slightly overwrites existing modifications in
dlgTriggerExport::slot_export() but is functionally the same there and also
in dlgTriggerEditor::slot_profileSaveAsAction() however in the latter there
was a (minor) logic error in the prior code that would misbehave if the
user manually specified a (valid, but old) file extension of ".trigger"
instead of ".xml" - the code should now work sensibly.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>