Improved map timing to be more human-friendly (#383)
We're not lab scientists here working with very detailed timing data, just people wanting to play a game - nine decimal level precision is unwarranted. In fact even reporting the time is unimportant, the only thing that matters is that it happens fast and when it's not fast, the user knows when it'll finish.