GMCP is now enabled by default on new profiles
GMCP has clearly eclipsed ATCP at this point by popularity and it has also been rock-solid for a couple of years now, so we can safely enable it by default and save people having to do the whole 'enable GMCP & reconnect' dance on a new profile.

Haven't got that much information about MSDP so it's not enabled yet.