Revert: undo previous changes to .clang-format file and reformat XMLimport
The modifications I made to the file used to control the action of
clang-format needed to be undone and then used for a re-run on the files
affected (the XMLimport class ones).

By comparison to the release_30 (the Mudlet 3.0 release) branch this
branch stores the same config file in the ./src sub-directory compared to
the parent directory that the other branch uses.  A revision to the
QMake project file ./src/src.pro was needed to display the file in the
Qt Creator IDE as I had put in the location for the release_30 version in
error in the previous commit.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>