Enhance: apply clang-formatting to XMLimport class
Following inter-developer discussion it was decided to apply the same
clang-format code formatting to the XMLimport class in this branch as
for the development branch - as this will increase the complexity of the
changes I felt there was little need to limit the scope of changes in this
branch to the minimum to provide the functional changes so have increased
the change-set to include some redundant code elimination steps already
done in the other branch commits to reduce the overall differences between
the changed files in the two branches.

I have also added the .clang-format file to the QMake project file so that
it shows up in the Qt Creator IDE - however the file is in the base
directory of the project in this branch's code whereas it is in the ./src/
sub-directory in the development branch at some point one branch should
be revised to use the same location for the file as the other one.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>