BugFix: fix missing icon from replay toolbar and a spammy replay qDebug
I spotted the missing icon recently and I recall it was something that had
already been addressed in the development branch (by AC in:
commit-9d52886268cfbb77498bedced2dc52ce6da1c815 )
but was not done for the release_30 branch for some reason.

There is also a qDebug() in (void)cTelnet::_loadreplay() that is in the
wrong place and which grossly fills up any debugging output as a result,
it seems I (or more likely Heiko) may have introduced it in:
commit-1a6106aa21cdfcb053a25eb0722de87325a24bbc
though AC did refactor it to use qDebug() when it previously used a
std::cout - unfortunately he did not spot that it would dump the entire
content of a replay chunk as many times as there were characters in that
chunk!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>