BugFix: attempt to repair introduced Bug.
In https://github.com/Mudlet/Mudlet/pull/372 (and a related PR for
release_30 branch) [bug 1670825](https://bugs.launchpad.net/bugs/1670825)
was found.  I think it might be that the change uncovered something that
was already at fault and this commit should address that.

This needs testing to see if it cures the matter.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>