Changes to make AppImage install working on Linux
Mostly getting Mudlet to look for files using an absolute path, because AppImage doesn't run set the current working directory for us.

Also updated LuaGlobal C++ searchpath to include absolute path. This is necessary for AppImage to work on Linux, which does not set the current working directory to be the mudlet executable.