Updated getMudletLuaDefaultPath+s to return a table of viable search locations
Added capability to return a LuaGlobal search path relative to the Mudlet binary - needed for AppImage on Linux which does not set the current working directory to the binary for us.

Renamed getMudletLuaDefaultPath() to getMudletLuaDefaultPaths() as it hasn't been out in a release yet and we need to return a list of values.