Revert: remove block on script compilation during profile start-up
The original behaviour was introduced in:
"Fix: made gui-kit functions profile-safe"
commit-b5a0b718ef3760e97152906230051f4c4bbcc20e
made by Heiko on 2009/05/03T16:54:34Z but with no further explanation as
to why. It only affected the compilation of lua code in the TScript class
but it now causes problems following a correction to:
TScript::setScript(QString &) which previously left the TScript::mOK_code
status flag in a state (true) that did not match reality.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>