Restore entry fields for command (down and up if the button is a push down one) to "Buttons" area of editor.
Shows or hides label and line-edit(s) for the above depending on whether
the displayed item is a button and is a push-down button - for menus all
are hidden as the process of clicking on a button to send a command is the
same action that menus use to pop-up their contents - and the
QPushButton::showMenu() method that is called to do the latter blocks until
the user makes a choice.  This also makes it inappropriate to send
something to the MUD server as part of popping up a menu item!

Also restore and enhance icon entry process so that it works in a user
friendly way - including buttons to set a new icon (which also displays the
current one) and clear an existing one (latter is only enabled if there IS
an icon set).

Fixes previous code errors that meant a toolbar contained in a module was
treated as a menu rather than a toolbar. Also fixes the colouration of
folder icons in the toolbar/menu/button editor so that brown folders are
modules, yellow folders are tool-bars and cyan ones are menus.

Module Master Folders are now NOT editable - which is consistent with their
implementation as they are virtual folders that are created by the module
import action and do not behave as a real TAction instance - deletion
requires removal .  Items can be
dragged into and out of them however and they may be deactivated and
activated to turn on or off their contents within the Editor for TActions
but not the other item types (so far)...!

Also:
* Adds the 48x48px Oxygen edit-delete.png icon for use on the clear icon
  button
* Adds the 48x48px Oxygen archive-remove.png icon to replace
  the edit-delete-shred.png icon for delete item on the editor toolbar

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>