Tweaks: revise formatting & revert some prior non-bugFixes
Reverting commenting out of TAlias::compile() as it isn't a "bugfix".

Added BOLD and BLUE formatting to Editor Error message for compilation of
TAlias regexp.

Reverting Lua related messages produced on the Central Debug Console during
run of TAlias::compileAll() so that the wording stays the same as that
produced for other Mudlet elements. Enhancements are intended in this area
but they are not BugFixes so should not be aimed at this branch.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>