(release 30) Linux generic installer fixes (#414)
* Changes to make AppImage install working on Linux

Mostly getting Mudlet to look for files using an absolute path, because AppImage doesn't run set the current working directory for us.

Also updated LuaGlobal C++ searchpath to include absolute path. This is necessary for AppImage to work on Linux, which does not set the current working directory to be the mudlet executable.

* Bugfix/Enhance: Load LuaGlobal.lua from LUA_DEFAULT_PATH

Add compile-time fallback to LuaGlobal.lua loading logic to attempt to load
from LUA_DEFAULT_PATH, defined in src/src.pro using the value from
LUA_DEFAULT_DIR.

For platforms that don't currently define LUA_DEFAULT_DIR in src/src.pro (e.g.
win32, macx), the loading logic will behave as though LUA_DEFAULT_DIR were
defined as "/", the root directory. This is accomplished by compile time
concatenation of LUA_DEFAULT_PATH and the string literal "/". If necessary,
LUA_DEFAULT_DIR can later be defined for additional platforms (win32, macx,
etc) in src/src.pro to provide a different compile-time fallback for that
platform.

(cherry picked from commit ee487888f219567635fb8334d980fd3fedd1f609)

* Updated getMudletLuaDefaultPath+s to return a table of viable search locations

Added capability to return a LuaGlobal search path relative to the Mudlet binary - needed for AppImage on Linux which does not set the current working directory to the binary for us.

Renamed getMudletLuaDefaultPath() to getMudletLuaDefaultPaths() as it hasn't been out in a release yet and we need to return a list of values.

* Changed load order to be 'closest to binary first'

This is the standard - search paths closest first and then those further away. Makes it easier to modify settings if need be.

* Add macOS lua path to getMudletLuaDefaultPaths

With this change, we should be able to get rid of the `run_mudlet` workaround for the macOS Mudlet.app package.

* Add Mac lua default path only when compiling there

To avoid confusion about search paths, the macOS Lua default path is only added when compiling for macOS.