Refactor: remove "duplicate" editor view constants from XMLimport class
It is not wise to duplicate a set of coded values when it is possible to
use the originals - this makes it safer should anything change to the
latter.  It does need to make the values `public` from the class where they
are defined but as they are static constants this should be safe to do.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>