BugFix: add Qt version conditional code for feature not supported in Qt<5.2
The current Qt Designer plugin in the Qt IDE is happy to enable features
that are not present in old Qt 5.x versions and if then asked to switch
them off does not remove them from the XML format .ui files it generates.

This commit fixes this problem that breaks compilation on Qt 5.x less than
5.2 for the clear "⌫" icon that can be used on QLineEdits to clear the
contents in one click in later versions. I had enabled this on the
"Command"/"Command (Down)" and "Command (Up)" entries on the Editor dialog
that handles Button/Menu features (some types of instances of the TAction
class).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>