Move setting of "build" information to own script
Due to the increasing complexity, I moved the setting of the build information to an own script.

This version will now check for PR builds and set the information accordingly, as well as remove the build information from tags completely.

It will change the src.pro only for qmake and the CMakeLists.txt only for cmake builds now.