Revise: use/store relative Button Icon pathFileNames
For more portable profiles this commit makes the location of icons used on
buttons to be stored/reported as relative to the profile home directory.

Something like this fix/enhancement is likely to also be needed for sound
files for triggers with audio file playback and lua playSoundFile(filename)
as they both use absolute files as well.

For backwards compatibility the data is saved in the profile save data in a
manner that means that prior versions of Mudlet will read the expected
absolute form but this and subsequent versions will read the relative form.

Also:
* increase the size of text on Icon setting buttons in Editor (were noted
as being too small!)
* SIGNAL/SLOT action of Icon setting buttons change to `clicked()` from
`pressed()` which means they do not stay "activated" during the execution
of the SLOT method connected to them (was noted in review).
* revised tooltip for Button/Menu/Toolbar "Name:" when a module/package
master folder is selected. (using text suggested during review with an
additon)
* added some `qDebug()` reports to trace behaviour of `TAction::mButtonState`
which is still not yet working as expected.
* refactored unused `TAction::execute(QStringList &)` and
misused `TAction::_execute(QStringList &)` to merge functionality AND
eliminate unused `QStringList &` argument.
* commented out unused `TAction::mIconPix` and made `TAction::mIcon`
private to hide internal manipulation of new relative/absolute pathFileName
code, revising:
 (QString) TAction::getIcon() ==>
     (QString) TAction::getIconPathFileName(
                                 const bool isToForceRelative=false) const
 (void) TAction::setIcon( QString & ) ==>
                       (void) TAction::setIconPathFileName(const QString &)

The PR that this is applied to is not yet complete as further work is
needed to fix up the puch down button behaviour - eventually the qDebug()
can be commented out/removed and then the series of commits should be
squashed at the merge point.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>