(release_30)BugFix_buttonCommandUpAndDownRedux_modulesWithToolBars
This is ANOTHER attempt to get the Command Up and Down for buttons restored
to Mudlet to get the release_30 branch ready for release.  Previous
attempts got bogged down in trying to do too much in one swell foop.

This provides the minimum IMHO to get the missing functionality back and
to fix a significant bug in the previous handling of toolbars that were
inserted into modules.

It is not everything that is wanted for the final release but represents
a functional package of changes that are hopefully possible to swallow in
one chunk.

It DOES NOT fix the behaviour of push-down buttons not sending the right
thing on every press - I am still investigating that urgently (I do have
some ideas); also the layout in the editor when some elements have
been necessarily been hidden is not perfect.  I have got a solution for
that but it is better wrapped up in a separate (series of) commit(s).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>