Refactor: remove redundant passing of global mudlet class pointer
The mudlet class is a singleton and there are means to get it thoughout
the application so it is pointless to pass it as an argument to a function
purely so it can be passed to another function.  This commit refactors the
code to use the global value directly where needed - which includes an area
of code that this series of commits affect.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>