BugFix: fix mButtonState behaviour for non-Push-Down "Buttons"
Revised: (void)TAction::execute() to always send the string associated with
the "down" command for a non-push down button. This fixes:
https://bugs.launchpad.net/mudlet/+bug/1673672 "Disabling button up still
keeps the command toggle". It also clearly sets TConsole::mButtonState to
1 if the button is up or 2 if it is down, as per:
http://wiki.mudlet.org/w/Manual:Lua_Functions#getButtonState .  Steps are
also taken to ensure a non-pushdown button is never left in a "down" state.

Refactored: (int)TAction::mButtonState ==> (bool)TAction::mButtonState
now true is used to indicate the button is in the pushed-down state - there
did appear to be some confusion about this in the past with 0,1 AND 2 being
used in various places.

BugFix: Disabled buttons are now hidden on floating toolbars as expected
{I do have future plans to allow the option for a button to be present but
"greyed-out" if disabled in the future...!}

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>