BugFix: Editor GUI visual improvements/behaviour
Renames some widgets to better identify them:
* (QWidget *)dlgActionMainArea::widget
                          ==> (QWidget *)dlgActionMainArea::widget_top
* (QWidget *)dlgActionMainArea::widget_4
                          ==> (QWidget *)dlgActionMainArea::widget_middle
* (QWidget *)dlgActionMainArea::widget_2
                          ==> (QGroupBox *)dlgActionMainArea::widget_bottom
* (QLabel *)dlgTimersMainArea::label_5
                                ==> (QLabel *)dlgTimersMainArea::label_name
* (QWidget *)dlgTriggersMainArea::widget_6
                             ==> (QWidget *)dlgTriggersMainArea::widget_top
* (QWidget *)dlgTriggersMainArea::widget
                          ==> (QWidget *)dlgTriggersMainArea::widget_middle
* (QWidget *)dlgTriggersMainArea::widget_3
                          ==> (QWidget *)dlgTriggersMainArea::widget_bottom
* (QLabel *)dlgTriggersMainArea::label_5
                          ==> (QLabel *)dlgTriggersMainArea::widget_command
* (QLabel *)dlgTriggersMainArea::label_4
                              ==> (QLabel *)dlgTriggersMainArea::label_name

Some "container" widgets have been removed from the .ui files used by
the editor (dlgTriggerEditor class):
* frame_rightBottom
* frame_rightTop
* mainArea
* popupArea
the code had to be revised to work on others, specifically:
  ==> mpSourceEditorArea

Revised all xxxx_main_area.ui files so that they all have "Name:" to
identify that aspect of the item that is being edited, and for all the
"direct string to send to MUD Server" entries to be "Command:" (obviously
the Action one has two under some conditions and none on others!

Revised all xxxx_main_area.ui files so that the ones that benefit from
extra space expand in the right places, for Triggers this is the middle
part which holds the list of trigger items and for Actions (Buttons /
Menus / Toolbars) this is the bottom widget which holds the "Optional CSS"
{which is now labelled "Style sheet:"}.  Keys, Aliases and Variables do not
benefit and Scripts only takes a little to expand the middle widget
containing a list of events that are handled.

The timers_main_area.ui was reformatted slightly to make better use of
vertical space (putting labels for each time element OVER the number
as opposed to a long single label to the right of all the digit displays.
The millisecond display was revised to show three digits (using format
"zzz") as they represent a decimal portion of a second rather than only
a single "0" when not otherwise set to anything.

Similarly the source editor area (which now gains an explicit "Lua code:")
label at the top can expand freely and has a minimum size set of around
200. Code has been added to change the text to "Value:" when Variables are
being edited.  This requires setting the text contained in the new:
* (QLabel) dlgSourceEditorArea::label_editor

The errors console - as it is a custom widget we construct in code also now
has a minimum size of 200 set but that has to be set in the
dlgTriggerEditor constructor.

Finally the system error messages widget has a minimum size set of 90 but
can now expand to 200 high.

All of these changes means that the splitter widget on the right hand side
of the editor is much more flexible and usable for a bigger range of
screen heights.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>